agents/lead_enrichment/requirements.txt
requests
python-dotenv
gspread
google-auth
google-auth-oauthlib
